feat: create client additional info zod object
