--- src/types/Media.ts
@@ -1,18 +1,0 @@
-import { z } from 'zod';
-
-// Payload Media type
-export const MediaSchema = z.object({
-  id: z.string(),
-  alt: z.string(),
-  updatedAt: z.string(),
-  createdAt: z.string(),
-  url: z.string().nullable().optional(),
-  thumbnailURL: z.string().nullable().optional(),
-  filename: z.string().nullable().optional(),
-  mimeType: z.string().nullable().optional(),
-  filesize: z.number().nullable().optional(),
-  width: z.number().nullable().optional(),
-  height: z.number().nullable().optional(),
-  focalX: z.number().nullable().optional(),
-  focalY: z.number().nullable().optional(),
-})
