+++ src/types/request-models/ProjectRequests.ts
@@ -1,21 +1,6 @@
 import { z } from 'zod'
 import { UserRole } from '../User'
-
-export const MediaSchema = z.object({
-  id: z.string(),
-  alt: z.string(),
-  updatedAt: z.string(),
-  createdAt: z.string(),
-  url: z.string().nullable().optional(),
-  thumbnailURL: z.string().nullable().optional(),
-  filename: z.string().nullable().optional(),
-  mimeType: z.string().nullable().optional(),
-  filesize: z.number().nullable().optional(),
-  width: z.number().nullable().optional(),
-  height: z.number().nullable().optional(),
-  focalX: z.number().nullable().optional(),
-  focalY: z.number().nullable().optional(),
-})
+import { MediaSchema } from '../Payload'
 
 export const UserSchema = z
   .object({
